+++ code/listSieve.c
@@ -2,6 +2,7 @@
 #include <stdio.h>
 #include <math.h>
 #include <string.h>
+#include <stdint.h>
 
 #include "listSieve.h"
 #include "common.h"
@@ -11,7 +12,7 @@
 #define k pow(2, (c*dim))
 
 // Function to perform the list sieve algorithm
-double listSieve(double** basis, double** L, double mu, long nSamples) {
+double listSieve(double** basis, double** L, double mu, int32_t nSamples) {
     double delta = 1 - (1 / dim);
     double* vReduced = NULL;
     double* diff;
@@ -32,7 +33,7 @@ double listSieve(double** basis, double** L, double mu, long nSamples) {
         int repeatFlag = isIn(vReduced, L, nSamples);
         // Check if v is in L
         if (repeatFlag == 0) {
-            for (long j = 0; j < nSamples; j++) {
+            for (int32_t j = 0; j < nSamples; j++) {
                 diff = vecDiff(vReduced, L[j]);
                 lenDiff = l2Norm(diff);
                 // Check if v can be reduced by a vector in L
